requirements.txt
Flask==3.1.1
gunicorn==23.0.0
PyJWT==2.10.1
requests==2.32.4
python-dotenv==1.0.1
